- camera management refactoring - codec management enhancements - code cleanup

+++ common/src/main/java/ru/coolsoft/common/Constants.java
@@ -5,6 +5,6 @@ public class Constants {
     public static final int SIZEOF_INT = 4;
     public static final int SIZEOF_LONG = 8;
 
-    public static final int CAMERA_UNAVAILABLE = 0;
-    public static final int CAMERA_AVAILABLE = 1;
+    public static final byte CAMERA_UNAVAILABLE = 0;
+    public static final byte CAMERA_AVAILABLE = 1;
 }
